refactor: 공용 상수를 constants/app.ts로 분리
