+++ apps/web/src/constants/app.ts
@@ -1,0 +1,3 @@
+export const COMPANY_NAME = '문어';
+export const TAGLINE = '문제없는 시공을 위한 어시스턴트';
+export const COPYRIGHT = `© ${new Date().getFullYear()} ${COMPANY_NAME}. All rights reserved.`;
